backend/requirements.txt
uWSGI==2.0.17.1
djangorestframework==3.8.2
django-cors-headers==2.4.0
Django==2.1.2
requests==2.19.1
psycopg2==2.7.5
pandas==0.23.4
sklearn
